Added DirectContentLinkBuilder to Siteimprove Adapter for recharge and other non CAE frontends.
